# docs/swagger.yaml
basePath: /api/
definitions:
  models.CreateEventRequest:
    properties:
      available_tickets:
        example: 20000
        type: integer
      date:
        example: "2024-12-31T20:00:00Z"
        type: string
      location:
        $ref: '#/definitions/models.CreateLocationRequest'
      name:
        example: Champions League Final
        type: string
      price:
        example: 99.99
        type: number
    type: object
  models.CreateLocationRequest:
    properties:
      address:
        example: 123 Main St
        type: string
      capacity:
        example: 50000
        type: integer
      country:
        example: USA
        type: string
      stadium:
        example: National Stadium
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      message:
        example: An error occurred
        type: string
    type: object
  models.EventResponse:
    properties:
      available_tickets:
        example: 15000
        type: integer
      date:
        example: "2024-12-31T20:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      location:
        $ref: '#/definitions/models.LocationResponse'
      name:
        example: Champions League Final
        type: string
      price:
        example: 99.99
        type: number
    type: object
  models.EventsResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/models.EventResponse'
        type: array
    type: object
  models.LocationResponse:
    properties:
      address:
        example: Wembley Park, London
        type: string
      capacity:
        example: 90000
        type: integer
      country:
        example: England
        type: string
      id:
        example: 101
        type: integer
      stadium:
        example: Wembley Stadium
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
  models.SuccessResponseCreate:
    properties:
      id:
        example: 1
        type: integer
      message:
        example: Object created successfully
        type: string
    type: object
...
  models.UpdateEventRequest:
    properties:
      available_tickets:
        example: 15000
        type: integer
      date:
        example: "2024-12-25T18:00:00Z"
        type: string
      location:
        $ref: '#/definitions/models.UpdateLocationRequest'
      name:
        example: Christmas Special
        type: string
      price:
        example: 49.99
        type: number
    type: object
  models.UpdateLocationRequest:
    properties:
      address:
        example: 456 Elm St
        type: string
      capacity:
        example: 75000
        type: integer
      country:
        example: Canada
        type: string
      stadium:
        example: Maple Leaf Stadium
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Simple API for managing ticket reservations.
  title: Ticket Reservation API
  version: "1.0"
paths:
  /events:
    get:
      description: Retrieve a list of all events with their details and locations.
      operationId: api.getEvents
      produces:
      - application/json
      responses:
        "200":
          description: List of events
          schema:
            $ref: '#/definitions/models.EventsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all events
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Parse the payload and create a new event with provided dataset.
      operationId: api.createEvent
      parameters:
      - description: Payload to create an event
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Event created successfully
          schema:
            $ref: '#/definitions/models.SuccessResponseCreate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new event.
      tags:
      - events
  /events/{id}:
    delete:
      description: Delete event by its ID.
      operationId: api.deleteEvent
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event deleted successfully
          schema:
            $ref: '#/definitions/models.SuccessResponseCreate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete an existing event
      tags:
      - events
    get:
      description: Retrieve an event with its details and location.
      operationId: api.getEventByID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event details
          schema:
            $ref: '#/definitions/models.EventResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get an event by ID
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Update event details based on the provided payload.
      operationId: api.updateEvent
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload to update an event
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Event updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update an existing event
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"